back/restapi/requirements.txt
amqp==5.1.1
asgiref==3.7.2
aws-secretsmanager-caching==1.1.1.5
billiard==4.1.0
boto3==1.34.70
celery==5.3.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.0
cryptography==41.0.3
depocs==1.0.1
Django==4.2.4
django-appconf==1.0.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-writable-nested==0.7.0
filetype==1.2.0
gprof2dot==2022.7.29
gunicorn==20.1.0
idna==3.4
kombu==5.3.2
mccabe==0.7.0
nulltype==2.3.1
ory-client==1.12.2
Pillow==10.0.0
plaid-python==14.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycodestyle==2.11.1
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
six==1.12.0
sqlparse==0.4.4
stripe==5.4.0
tzdata==2023.3
ua-parser==0.18.0
urllib3==2.0.3
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.7
